requirements.txt
requests>=2.31.0,<3
tweepy==4.14.0
python-dotenv==1.0.0
tqdm==4.66.1
backoff==2.2.1